+++ server/rkwebscrd.py
@@ -847,7 +847,7 @@ class NativeEncoder:
 
 
 class RequestHandler(BaseHTTPRequestHandler):
-    server_version = "rkwebscr/0.4.2"
+    server_version = "rkwebscr/0.4.3"
 
     @property
     def app(self) -> "Application":